Renamed tracing subscriber logger
