+++ crates/burn-train/src/learner/mod.rs
@@ -7,6 +7,7 @@ mod regression;
 mod step;
 mod summary;
 mod train_val;
+mod application_logger;
 
 pub use base::*;
 pub use builder::*;
@@ -18,3 +19,4 @@ pub use step::*;
 pub use summary::*;
 pub use train::*;
 pub use train_val::*;
+pub use application_logger::*;
