+++ crates/burn-train/src/logger/mod.rs
@@ -3,11 +3,9 @@ mod base;
 mod file;
 mod in_memory;
 mod metric;
-mod tracing;
 
 pub use async_logger::*;
 pub use base::*;
 pub use file::*;
 pub use in_memory::*;
 pub use metric::*;
-pub use tracing::*;
