+++ examples/custom-renderer/src/lib.rs
@@ -76,7 +76,7 @@ pub fn run<B: AutodiffBackend>(device: B::Device) {
         .devices(vec![device])
         .num_epochs(config.num_epochs)
         .renderer(CustomRenderer {})
-        .with_logger(None);
+        .with_application_logger(None);
     // can be used to interrupt training
     let _interrupter = builder.interrupter();
